Updated!
The only way I could clean up the added test app was to give it its own package. However, doing that made access to the icon file go away, so I added a small method to the RunnableTest so that I could pull the URL to the icon file. Seems kinda hokey to me but I could not find  way to get to that icon file once I was in the new package context.

I could not find any hanging or incomplete javadoc comments ... I'm not sure where you saw that, and its possible that I unknowingly got rid of it.

I cleaned up my clutter this time :-)