Minor updates and fixes
Removed the thread that adds menuItems and rolled the isUnique test into the EventQue thread.

Changed the labels of the two default MenuItems for consistency.

Changed the default top menu's add method to an Insert at index 0 so that it is always at the top of the menu.

added the insertMenuItem method for added functionality.

Moved all private methods to the bottom of the class.