Re-structured the Builder class so that it now conforms to "proper" Builder style. The Builder class now only accepts options that are desired, then it returns a new instance of FXTrayIcon via a private constructor that only accepts an instance of the Builder class as the argument. That constructor then first calls FXTrayIcons main constructor then it proceeds to configure all of the options that were set in the Builder class, including optional MenuItems and other desired behaviors.
Configured default icon size to be set based on operating system both in Builder and parent constructors - if there is no value passed into the library for icon dimensions.

Added an overloaded constructor where only the parentStage is required. FXTrayIcon will use a default graphic for the trayIcon when this constructor is used. This feature existed already in the Builder class, I simply extended it to the parent constructors.

Also tested every runnable test, updated some to use the builder class and replaced the old purple chain icon with the modern looking icon.

Fixed the completely picky code checker "errors" that is assigned to FXTrayIcon Git repository --- where it specifically failed the pull request version of code.